requirements.txt
allure-pytest
requests
pytest==7.3.1